Move texture instructions into a separate file
